Initial step of extracting AbstractBuildOperationRunner
